shows download dialog when downloading attachments

+++ app/src/main/java/com/example/egercomms/services/announcement/FileDownloadService.java
@@ -1,8 +1,14 @@
 package com.example.egercomms.services.announcement;
 
 import android.app.IntentService;
+import android.app.NotificationChannel;
+import android.app.NotificationManager;
+import android.content.Context;
 import android.content.Intent;
+import android.graphics.Color;
+import android.os.Build;
 import android.os.Environment;
+import android.support.v4.app.NotificationCompat;
 import android.support.v4.content.LocalBroadcastManager;
 import android.util.Log;
 
@@ -30,6 +36,8 @@ public class FileDownloadService extends IntentService {
     private String message;
     private DataHandler dataHandler = DataHandler.getInstance();
     private boolean permissionGranted;
+    private NotificationCompat.Builder notificationBuilder;
+    private NotificationManager notificationManager;
 
     public FileDownloadService() {
         super("FileDownloadService");
@@ -42,6 +50,25 @@ public class FileDownloadService extends IntentService {
         String url = builder.append(BASE_URL).append(path).toString();
         FileDownloadWebService webService = FileDownloadWebService.retrofit.create(FileDownloadWebService.class);
         if (permissionGranted) {
+            notificationManager = (NotificationManager) getSystemService(Context.NOTIFICATION_SERVICE);
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+                NotificationChannel notificationChannel = new NotificationChannel("id", "an", NotificationManager.IMPORTANCE_LOW);
+
+                notificationChannel.setDescription("no sound");
+                notificationChannel.setSound(null, null);
+                notificationChannel.enableLights(false);
+                notificationChannel.setLightColor(Color.BLUE);
+                notificationChannel.enableVibration(false);
+                notificationManager.createNotificationChannel(notificationChannel);
+            }
+            notificationBuilder = new NotificationCompat.Builder(this, "id")
+                    .setSmallIcon(android.R.drawable.stat_sys_download)
+                    .setContentTitle("Download")
+                    .setContentText("Downloading File")
+                    .setDefaults(0)
+                    .setAutoCancel(true);
+            notificationManager.notify(0, notificationBuilder.build());
+
             Call<ResponseBody> call = webService.downloadAttachments(url);
             call.enqueue(new Callback<ResponseBody>() {
                 @Override
@@ -51,7 +78,11 @@ public class FileDownloadService extends IntentService {
 
                         boolean writtenToDisk = writeResponseBodyToDisk(response.body());
 
+                        if(writtenToDisk){
                         message = "File Downloaded to Egercomms/downloads";
+                        }else{
+                            message = "Error downloading file. Check if it exists";
+                        }
 
                         Log.d(TAG, "file download was a success? " + writtenToDisk);
 
@@ -71,7 +102,7 @@ public class FileDownloadService extends IntentService {
                     returnResults(message);
                 }
             });
-        }else{
+        } else {
             returnResults("You must grant file permissions to download files");
         }
     }
@@ -86,56 +117,75 @@ public class FileDownloadService extends IntentService {
 
     private boolean writeResponseBodyToDisk(ResponseBody body) {
         try {
-            String folderPath = "Egercomms"+File.separator+"downloads";
+            String folderPath = "Egercomms" + File.separator + "downloads";
             File folder = new File(Environment.getExternalStorageDirectory(), folderPath);
-            if(!folder.exists()){
+            if (!folder.exists()) {
                 folder.mkdirs();
             }
-            File file = new File(Environment.getExternalStorageDirectory(), folderPath+File.separator + getFileName(path));
+            File file = new File(Environment.getExternalStorageDirectory(), folderPath + File.separator + getFileName(path));
+            if (!file.exists()) {
+                InputStream inputStream = null;
+                OutputStream outputStream = null;
 
-            InputStream inputStream = null;
-            OutputStream outputStream = null;
+                try {
+                    byte[] fileReader = new byte[4096];
+                    int count;
 
-            try {
-                byte[] fileReader = new byte[4096];
+                    long fileSize = body.contentLength();
+                    long fileSizeDownloaded = 0;
+                    boolean downloadComplete = false;
+                    inputStream = body.byteStream();
+                    outputStream = new FileOutputStream(file);
 
-                long fileSize = body.contentLength();
-                long fileSizeDownloaded = 0;
 
-                inputStream = body.byteStream();
-                outputStream = new FileOutputStream(file);
+                    while ((count = inputStream.read(fileReader)) != -1) {
+                        fileSizeDownloaded += count;
+                        int progress = (int) ((double) (fileSizeDownloaded * 100) / (double) fileSize);
 
-                while (true) {
-                    int read = inputStream.read(fileReader);
-
-                    if (read == -1) {
-                        break;
+                        updateNotification(progress);
+                        outputStream.write(fileReader, 0, count);
+                        downloadComplete = true;
                     }
+                    onDownloadComplete(downloadComplete);
+                    outputStream.flush();
 
-                    outputStream.write(fileReader, 0, read);
-
-                    fileSizeDownloaded += read;
-
-                    Log.d(TAG, "file download: " + fileSizeDownloaded + " of " + fileSize);
+                    return true;
+                } catch (IOException e) {
+                    return false;
+                } finally {
+                    if (inputStream != null) {
+                        inputStream.close();
+                    }
+                    if (outputStream != null) {
+                        outputStream.close();
+                    }
                 }
-
-                outputStream.flush();
-
-                return true;
-            } catch (IOException e) {
+            }else{
+                notificationManager.cancel(0);
                 return false;
-            } finally {
-                if (inputStream != null) {
-                    inputStream.close();
-                }
-
-                if (outputStream != null) {
-                    outputStream.close();
-                }
             }
         } catch (IOException e) {
             return false;
         }
+    }
+
+    private void updateNotification(int currentProgress) {
+        notificationBuilder.setProgress(100, currentProgress, false);
+        notificationBuilder.setContentText("Downloaded: " + currentProgress + "%");
+        notificationManager.notify(0, notificationBuilder.build());
+    }
+
+    private void onDownloadComplete(boolean downloadComplete) {
+        notificationManager.cancel(0);
+        notificationBuilder.setProgress(0, 0, false);
+        notificationBuilder.setContentText("File Download Complete");
+        notificationManager.notify(0, notificationBuilder.build());
+
+    }
+
+    @Override
+    public void onTaskRemoved(Intent rootIntent) {
+        notificationManager.cancel(0);
     }
 
     private String getFileName(String path) {
